More progress

+++ include/Texture.hpp
@@ -13,8 +13,8 @@ private:
 
 public:
   Texture(std::string filename, SDL_Renderer *renderer);
-  double get_width() const;
-  double get_height() const;
+  unsigned get_width() const;
+  unsigned get_height() const;
   void render(const SDL_FRect *dst) const;
   void render(const SDL_FRect *src, double x, double y, double w,
               double h) const;
